requirements.txt
pycryptodome
requests==2.30.0
urllib3==1.26.6